requirements.txt
Django==4.2.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
Pillow==10.0.0
requests==2.31.0
python-dateutil==2.8.2
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.7.0
